requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.1
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==25.0
pillow==11.3.0
pygame==2.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0
zipp==3.23.0
